% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.R
\name{add_intercept_column}
\alias{add_intercept_column}
\title{Add an intercept column to \code{x}}
\usage{
add_intercept_column(x, name = "(Intercept)")
}
\arguments{
\item{x}{A data frame or matrix.}

\item{name}{The name for the intercept column. Defaults to \code{"(Intercept)"},
which is the same name that \code{\link[stats:lm]{stats::lm()}} uses.}
}
\description{
This function adds an integer column of \code{1}'s to \code{x}.
}
\details{
If a column named \code{name} already exists in \code{x}, then \code{x} is returned
unchanged and a warning is issued.
}
\examples{
add_intercept_column(mtcars)

add_intercept_column(mtcars, "intercept")

add_intercept_column(as.matrix(mtcars))

}
