% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-pred.R
\name{check_quantile_levels}
\alias{check_quantile_levels}
\title{Check levels of quantiles}
\usage{
check_quantile_levels(levels, call = rlang::caller_env())
}
\arguments{
\item{levels}{The quantile levels.}

\item{arg, call}{Inputs to use to write error messages}
}
\value{
Invisible \code{TRUE}
}
\description{
Check levels of quantiles
}
\details{
Checks the levels for their data type, range, uniqueness, and missingness.
}
\keyword{internal}
