% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr_one_hot.R
\name{contr_one_hot}
\alias{contr_one_hot}
\title{Contrast function for one-hot encodings}
\usage{
contr_one_hot(n, contrasts = TRUE, sparse = FALSE)
}
\arguments{
\item{n}{A vector of factor levels or the number of unique levels.}

\item{contrasts}{Should contrasts be computed. Only a value of \code{TRUE} is used
(this argument is included for S comparability).}

\item{sparse}{A logical for a sparse matrices. Only non-sparse formats are
currently implemented.}
}
\value{
A diagonal matrix that is \code{n}-by-\code{n}.
}
\description{
This contrast function produces a model matrix that has indicator columns for
each level of each factor.
}
\examples{
dat <-
  data.frame(
    x = 1:12,
    y = factor(rep(letters[1:3], each = 4)),
    z = factor(rep(LETTERS[1:2], 6))
  )

## -----------------------------------------------------------------------------

# No intercept does not produce all columns for all factors:
model.matrix(x ~ y + z + 0, data = dat)

## -----------------------------------------------------------------------------

# Save current contrast to reset at bottom
old_contr <- options("contrasts")$contrasts
new_contr <- old_contr
new_contr["unordered"] <- "contr_one_hot"
options(contrasts = new_contr)

## -----------------------------------------------------------------------------

model.matrix(x ~ y + z, data = dat)

## -----------------------------------------------------------------------------

options(contrasts = old_contr)

}
