% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recompose.R
\name{convert_matrix}
\alias{convert_matrix}
\title{Convert tibble to matrix}
\usage{
convert_matrix(x, sparse = TRUE)
}
\arguments{
\item{x}{A tibble.}

\item{sparse}{Logical, denoting if result should be a sparse matrix.}
}
\value{
A matrix or dgCMatrix.
}
\description{
Convert tibble to matrix
}
\examples{
convert_matrix(mtcars, sparse = TRUE)

convert_matrix(mtcars, sparse = FALSE)
}
