% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{default_preprocessor}
\alias{default_preprocessor}
\title{Create a default preprocessor}
\usage{
default_preprocessor()
}
\description{
A default preprocessor is a function that takes in \code{new_data},
\code{intercept}, and \code{type}. It performs some basic preprocessing on \code{new_data}
to prepare it for ingestion into a model. A default preprocessor is used
in matrix and data frame methods of a model (as opposed to a recipe or
formula method which performs preprocessing for you).
}
\details{
The preprocessor function returned from \code{default_preprocessor()} will do two
things:
\itemize{
\item Call \code{\link[=retype]{retype()}} with \code{type} to coerce \code{new_data} to a specific type.
\item Call \code{\link[=add_intercept_column]{add_intercept_column()}} with \code{intercept} to add an intercept to
\code{new_data} if required.
}

The returned function that \code{default_preprocessor()} creates has 3 arguments:
\itemize{
\item \code{new_data}: The data to preprocess.
\item \code{intercept}: A logical. Passed on to \code{add_intercept_column()}.
\item \code{type}: A single character. Passed on to \code{retype()}.
}
}
