% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce.R
\name{enforce_new_data_novel_levels}
\alias{enforce_new_data_novel_levels}
\title{Check for new factor levels}
\usage{
enforce_new_data_novel_levels(new_data, original_levels)
}
\arguments{
\item{new_data}{A data frame of new predictors and maybe outcomes.}

\item{original_levels}{A named list of the original levels of either the
outcomes or predictors. These are stored in the preprocessor slots
\code{$predictors$levels} and \code{$outcomes$levels}.}
}
\description{
A factor column of \code{new_data} might have \emph{new} factor levels when compared
to the original levels used in training. These new levels are undefined, and
are caught when preprocessing the \code{new_data} and are converted to \code{NA} with
a warning.
}
\keyword{internal}
