% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{extract_info}
\alias{extract_info}
\title{Extract class and type information}
\usage{
extract_info(x)
}
\arguments{
\item{x}{A data frame or matrix.}
}
\value{
A named list with 3 elements: \code{names}, \code{classes}, and \code{levels}.
}
\description{
\code{extract_info()} extracts the following pieces of information from \code{x}:
\itemize{
\item \code{names}: Using \code{colnames()}.
\item \code{classes}: Use \code{\link[=get_data_classes]{get_data_classes()}}.
\item \code{levels}: Using \code{\link[=get_levels]{get_levels()}}.
}
}
\details{
\code{extract_info()} is useful when creating new preprocessing \code{engine}. It
extracts the required information that will be used by the validation checks
at prediction time.
}
\examples{

str(hardhat:::extract_info(iris))

}
\keyword{internal}
