% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-outcome.R
\name{extract_outcome}
\alias{extract_outcome}
\title{Extract the outcome}
\usage{
extract_outcome(y, mode, variateness)
}
\arguments{
\item{y}{The outcome.}

\item{mode}{Either \code{"classification"} or \code{"regression"}.}

\item{variateness}{Either \code{"univariate"} or \code{"multivariate"} depending
on whether or not you have >1 outcome per observation.}
}
\description{
Most of the time, the input to a model should be flexible enough to capture
a number of different input types from the user. \code{extract_outcome()} captures
that flexibility for you, and converts user input of various types into a
common output type based on the \code{mode} and \code{variateness} of your model.
}
\section{Univariate Regression}{


\emph{Returns:}
\itemize{
\item A single numeric vector
}

\emph{Accepts:}
\itemize{
\item A numeric vector
\item A data frame with 1 numeric column
\item A numeric matrix with 1 column
}
}

\section{Multivariate Regression}{


\emph{Returns:}
\itemize{
\item A numeric matrix with as many columns as there are outcomes
}

\emph{Accepts:}
\itemize{
\item A data frame with >1 numeric columns
\item A numeric matrix with >1 columns
}
}

\section{Univariate Classification}{


\emph{Returns:}
\itemize{
\item A factor
}

\emph{Accepts:}
\itemize{
\item A factor
\item A data frame with 1 factor column
}
}

\section{Multivariate Classification}{


\emph{Returns:}
\itemize{
\item A data frame with as many factor columns as there are outcomes
}

\emph{Accepts:}
\itemize{
\item A data frame with >1 factor columns
}
}

