% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-levels.R
\name{extract_outcome_levels}
\alias{extract_outcome_levels}
\title{Extract the levels from a factor outcome}
\usage{
extract_outcome_levels(y, mode, variateness)
}
\arguments{
\item{y}{The outcome. A factor or data frame of factors.}

\item{mode}{A character. Generally either \code{"classification"}
or \code{"regression"}.}

\item{variateness}{A character. One of \code{"univariate"} or \code{"multivariate"}.}
}
\value{
If \code{variateness == "univariate"}, a single character vector for the
levels of \code{y}.

If \code{variateness == "multivariate"}, a named list of character vectors
for the levels of each factor column of \code{y}.
}
\description{
If \code{y} is a factor, then \code{extract_outcome_levels()} will return the levels of
the factor as a character vector. If \code{y} is a data frame of factors
a named list of character vectors will be returned.
}
\examples{

y <- as.factor(letters[1:5])

extract_outcome_levels(y, "classification", "univariate")

extract_outcome_levels(as.data.frame(y), "classification", "univariate")

z <- as.factor(letters[6:10])
y_z_df <- data.frame(y = y, z = z)

extract_outcome_levels(y_z_df, "classification", "multivariate")

}
