% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge.R
\name{forge.terms_preprocessor}
\alias{forge.terms_preprocessor}
\title{Forge - Formula Method}
\usage{
\method{forge}{terms_preprocessor}(preprocessor, new_data,
  outcomes = FALSE, ...)
}
\value{
A named list with 3 elements:
\itemize{
\item \code{predictors}: A tibble containing the preprocessed
\code{new_data} predictors.
\item \code{outcomes}: If \code{outcomes = TRUE}, a tibble containing the preprocessed
\code{new_data} outcomes. Otherwise, \code{NULL}.
\item \code{offset}: If the \code{preprocessor} was a \code{"terms_preprocessor"}, and offsets
were specified in the formula, this is a tibble containing the preprocessed
offsets. Otherwise, \code{NULL}.
}
}
\description{
If a formula engine was used, this is a data frame
that is the result of extracting the outcome columns from
\code{\link[stats:model.frame]{stats::model.frame()}}
}
