% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge.R
\name{forge.recipes_preprocessor}
\alias{forge.recipes_preprocessor}
\title{Forge - Recipes Method}
\usage{
\method{forge}{recipes_preprocessor}(preprocessor, new_data,
  outcomes = FALSE, ...)
}
\arguments{
\item{preprocessor}{A \code{"recipes_preprocessor"}.}

\item{new_data}{A data frame or matrix to preprocess.}

\item{outcomes}{A logical. Should the outcomes be processed and returned
as well?}

\item{...}{Not currently used.}
}
\value{
A named list with 3 elements:
\itemize{
\item \code{predictors}: A tibble containing the preprocessed
\code{new_data} predictors.
\item \code{outcomes}: If \code{outcomes = TRUE}, a tibble containing the preprocessed
\code{new_data} outcomes. Otherwise, \code{NULL}.
\item \code{offset}: If the \code{preprocessor} was a \code{"terms_preprocessor"}, and offsets
were specified in the formula, this is a tibble containing the preprocessed
offsets. Otherwise, \code{NULL}.
}
}
\description{
If a recipe was used, this is a data.frame that
is the result of calling \code{\link[recipes:bake]{recipes::bake()}} with \code{\link[recipes:all_outcomes]{recipes::all_outcomes()}}
specified.
}
