% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_data_classes}
\alias{get_data_classes}
\title{Extract data classes from a data frame or matrix}
\usage{
get_data_classes(x)
}
\arguments{
\item{x}{A data frame or matrix.}
}
\value{
A named list. The names are the column names of \code{x} and the values are
character vectors containing the class of that column.
}
\description{
When predicting from a model, it is often important for the \code{new_data} to
have the same classes as the original data used to fit the model.
\code{get_data_classes()} extracts the classes from the original training data.
}
\details{
\code{get_data_classes()} is inspired by \code{\link[stats:.MFclass]{stats::.MFclass()}}. It returns the
exact classes of the data frame columns with the one exception that integer
and double columns are both returned as \code{"numeric"}. This is to ensure
that if an integer column is used when fitting, an error is not thrown if
a double column is used at prediction time. It also does not allow for
any columns of \code{x} to be multivariate themselves (no data frame columns,
or matrix columns).

The classes returned by \code{get_data_classes()} are stored in the \code{engine}
objects. This function is called by \code{\link[=mold]{mold()}}.
}
\examples{
get_data_classes(iris)

get_data_classes(as.matrix(mtcars))

# Unlike .MFclass(), the full class
# vector is returned
x <- data.frame(col = ordered(c("a", "b")))

.MFclass(x$col)

get_data_classes(x)

}
