% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{get_indicators}
\alias{get_indicators}
\title{Get indicator matrix from factor}
\usage{
get_indicators(x)
}
\arguments{
\item{x}{A factor without missing values.}
}
\value{
A integer matrix with the same number of columns as levels of \code{x},
and the same number of rows as the number of elements in \code{x}.
}
\description{
Takes a factor vector and construct the corresponding indicator matrix, also
known as a one-hot encoding matrix.
}
\examples{
get_indicators(factor(letters))

get_indicators(factor(letters[1:2], levels = letters))

set.seed(1234)
get_indicators(factor(sample(letters[1:4], 10, TRUE)))
}
