% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat-example-data.R
\docType{data}
\name{hardhat-example-data}
\alias{hardhat-example-data}
\alias{example_train}
\alias{example_test}
\title{Example data for hardhat}
\value{
\item{example_train,example_test}{tibbles}
}
\description{
Example data for hardhat
}
\details{
Data objects for a training and test set with the same variables:
three numeric and two factor columns.
}
\examples{
data("hardhat-example-data")
}
\keyword{datasets}
