% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{hardhat-extract}
\alias{hardhat-extract}
\alias{extract_workflow}
\alias{extract_recipe}
\alias{extract_parsnip_spec}
\alias{extract_fit_parsnip}
\alias{extract_fit_engine}
\title{Generics for object extraction}
\usage{
extract_workflow(x, ...)

extract_recipe(x, ...)

extract_parsnip_spec(x, ...)

extract_fit_parsnip(x, ...)

extract_fit_engine(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Extra arguments passed on to methods.}
}
\description{
These generics are used to extract elements from various model
objects. Methods are defined in other packages, such as tune,
workflows, and workflowsets, but the returned object is always the same.
\itemize{
\item \code{extract_workflow()} returns a workflow, possibly fit.
\item \code{extract_recipe()} returns a recipe, possibly prepped.
\item \code{extract_parsnip_spec()} returns a parsnip model specification.
\item \code{extract_fit_parsnip()} returns a parsnip model fit.
\item \code{extract_fit_engine()} returns the engine specific fit embedded within
a parsnip model fit. For example, when using \code{\link[parsnip:linear_reg]{parsnip::linear_reg()}}
with the \code{"lm"} engine, this would return the underlying lm object.
}
}
\examples{
# See packages where methods are defined for examples.
}
