% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat-package.R
\docType{package}
\name{hardhat-package}
\alias{hardhat}
\alias{hardhat-package}
\title{hardhat: Construct Modeling Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Building modeling packages is hard. A large amount of effort generally goes into providing an implementation for a new method that is efficient, fast, and correct, but often less emphasis is put on the user interface. A good interface requires specialized knowledge about S3 methods and formulas, which the average package developer might not have. The goal of 'hardhat' is to reduce the burden around building new modeling packages by providing functionality for preprocessing, predicting, and validating input.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/hardhat}
  \item \url{https://hardhat.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/hardhat/issues}
}

}
\author{
\strong{Maintainer}: Hannah Frick \email{hannah@posit.co} (\href{https://orcid.org/0000-0002-6049-5258}{ORCID})

Authors:
\itemize{
  \item Hannah Frick \email{hannah@posit.co} (\href{https://orcid.org/0000-0002-6049-5258}{ORCID})
  \item Davis Vaughan \email{davis@posit.co}
  \item Max Kuhn \email{max@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
