% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat_examples.R
\docType{data}
\name{hardhat_examples}
\alias{hardhat_examples}
\alias{example_train}
\alias{example_test}
\title{Examples for hardhat}
\value{
\item{example_train,example_test}{tibbles}
}
\description{
Examples for hardhat
}
\details{
Data objects for a training and test set with the same variables:
three numeric and two factor columns.
}
\examples{
data(hardhat_examples)
}
\keyword{datasets}
