% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{is_blueprint}
\alias{is_blueprint}
\title{Is \code{x} a preprocessing blueprint?}
\usage{
is_blueprint(x)
}
\arguments{
\item{x}{An object.}
}
\description{
\code{is_blueprint()} checks if \code{x} inherits from \code{"hardhat_blueprint"}.
}
\examples{
is_blueprint(default_xy_blueprint())

}
