% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-weights.R
\name{is_case_weights}
\alias{is_case_weights}
\title{Is \code{x} a case weights vector?}
\usage{
is_case_weights(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
\code{is_case_weights()} checks if \code{x} inherits from \code{"hardhat_case_weights"}.
}
\examples{
is_case_weights(1)
is_case_weights(frequency_weights(1))
}
