% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{is_engine}
\alias{is_engine}
\title{Is \code{x} a preprocessing engine?}
\usage{
is_engine(x)
}
\arguments{
\item{x}{An object.}
}
\description{
\code{is_engine()} checks if \code{x} inherits from \code{"hardhat_engine"}.
}
\examples{
is_engine(default_xy_engine())

}
