% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{is_preprocessor}
\alias{is_preprocessor}
\title{Is \code{x} a valid preprocessor?}
\usage{
is_preprocessor(x)
}
\arguments{
\item{x}{An object.}
}
\description{
This function checks to see if \code{x} is a valid \code{"preprocessor"}.
}
\examples{

x <- mold(Species ~ Sepal.Width, iris)
is_preprocessor(x$preprocessor)

}
