% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{is_preprocessor_engine}
\alias{is_preprocessor_engine}
\title{Is \code{x} a valid preprocessor engine?}
\usage{
is_preprocessor_engine(x)
}
\arguments{
\item{x}{An object.}
}
\description{
This function checks to see if \code{x} is a valid preprocessing engine. This
can be a \code{recipe}, a \code{terms} object, or a \code{default_preprocessor_engine}.
}
\keyword{internal}
