% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-pred.R
\name{is_quantile_pred}
\alias{is_quantile_pred}
\title{Test if an object is a \code{quantile_pred}}
\usage{
is_quantile_pred(x)
}
\arguments{
\item{x}{an object}
}
\value{
A boolean.
}
\description{
Test if an object is a \code{quantile_pred}
}
\examples{
qp <- quantile_pred(matrix(rnorm(10), nrow = 2), 1:5 / 6)
is_quantile_pred(qp)
is_quantile_pred(1:5)
}
