% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-formula-default.R,
%   R/engine-recipe-default.R, R/engine-xy-default.R
\name{new_default_formula_engine}
\alias{new_default_formula_engine}
\alias{new_default_recipe_engine}
\alias{new-default-engine}
\alias{new_default_xy_engine}
\title{Create a new default engine}
\usage{
new_default_formula_engine(mold, forge, intercept = FALSE, info = NULL,
  formula = NULL, indicators = TRUE, terms = list(predictors = NULL,
  outcomes = NULL), ..., subclass = character())

new_default_recipe_engine(mold, forge, intercept = FALSE, info = NULL,
  recipe = NULL, ..., subclass = character())

new_default_xy_engine(mold, forge, intercept = FALSE, info = NULL, ...,
  subclass = character())
}
\arguments{
\item{mold}{A named list with two elements, \code{clean} and \code{process}, see
the \code{\link[=new_engine]{new_engine()}} section, Mold Functions, for details.}

\item{forge}{A named list with two elements, \code{clean} and \code{process}, see
the \code{\link[=new_engine]{new_engine()}} section, Forge Functions, for details.}

\item{intercept}{A logical. Should an intercept be included in the
processed data? This information is used by the \code{process} function
in the \code{mold} and \code{forge} function list.}

\item{info}{Either \code{NULL}, or a named list with 2 elements, \code{predictors}
and \code{outcomes}. \code{info} is generated automatically at \code{\link[=mold]{mold()}} time and
is used to validate \code{new_data} at prediction time. At \code{\link[=mold]{mold()}} time, the
information found in \code{engine$mold$process()$info} is used to set \code{info}
for the \code{engine}.}

\item{formula}{Either \code{NULL}, or a formula that specifies how the
predictors and outcomes should be preprocessed. This argument is set
automatically at \code{\link[=mold]{mold()}} time.}

\item{indicators}{A logical. Should factors be expanded into dummy variables?}

\item{terms}{A named list of two elements, \code{predictors} and \code{outcomes}. Both
elements are \code{terms} objects that describe the terms for the outcomes and
predictors separately. This argument is set automatically at \code{\link[=mold]{mold()}} time.}

\item{...}{Name-value pairs for additional elements of engines that
subclass this engine.}

\item{subclass}{A character vector. The subclasses of this engine.}

\item{recipe}{Either \code{NULL}, or an unprepped recipe. This argument is set
automatically at \code{\link[=mold]{mold()}} time.}
}
\description{
This page contains the constructors for the default engines. They can be
extended if you want to add extra behavior on top of what the default
engines already do, but generally you will extend the non-default versions
of the constructors found in the documentation for \code{\link[=new_engine]{new_engine()}}.
}
