% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{new_default_preprocessor_engine}
\alias{new_default_preprocessor_engine}
\title{Create a default preprocessor engine}
\usage{
new_default_preprocessor_engine()
}
\description{
A default preprocessor engine is a function that performs some basic
preprocessing on \code{new_data} to mold it for ingestion into a model. To
control how the engine preprocesses \code{new_data} the \code{intercept} argument can
be set, which is passed down to the created engine function.
A default preprocessor engine is used in matrix and data frame methods of a
model (as opposed to a recipe or formula method which performs
preprocessing for you). This preprocessing is performed on both the training
and testing data sets.
}
\details{
The preprocessor function returned from \code{new_default_preprocessor_engine()}
will only do one thing:
\itemize{
\item Call \code{add_intercept_column()} with \code{intercept} to add an intercept to
\code{new_data} if required.
}

The returned function that \code{new_default_preprocessor_engine()} creates
has 2 arguments:
\itemize{
\item \code{new_data}: The data to preprocess.
\item \code{intercept}: A logical. Passed on to \code{add_intercept_column()}.
}
}
\keyword{internal}
