% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{new_engine}
\alias{new_engine}
\title{Create a new preprocessing engine}
\usage{
new_engine(mold, forge, intercept = FALSE, info = NULL, ...,
  subclass = character())
}
\arguments{
\item{mold}{A named list with two elements:
\itemize{
\item \code{clean}: A function that performs initial cleaning of the user's input
data to be used in the model.
\item \code{process}: A function that performs the actual preprocessing of the data.
}

Both \code{engine$mold$clean()} and \code{engine$mold$process()} will be called,
in order, from \code{\link[=mold]{mold()}} and will be passed the \code{engine} itself, and the
data (\code{x} and \code{y} for the xy method, but otherwise \code{data}).}

\item{forge}{A named list with two elements:
\itemize{
\item \code{clean}: A function that performs initial cleaning of the \code{new_data} that
is typically used in generating predictions.
\item \code{process}: A function that performs the preprocessing of the \code{new_data}.
}

Both \code{engine$forge$clean()} and \code{engine$forge$process()} will be called,
in order, from \code{\link[=mold]{mold()}} and will be passed the \code{engine} itself,
and the \code{new_data}.}

\item{intercept}{A logical. Should an intercept be included in the
processed data? This information is used by the \code{process} function
in the \code{mold} and \code{forge} function list.}

\item{info}{Either \code{NULL}, or a named list with 2 elements:
\itemize{
\item \code{predictors}: A named list with the following 3 elements:
\itemize{
\item \code{names}: A character vector of the original predictor column names.
\item \code{classes}: A named list of the original predictor classes, or \code{NULL}.
\item \code{levels}: A named list of the original predictor levels for any factor
columns, or \code{NULL}.
}
\item \code{outcomes}: A named list with the following 3 elements:
\itemize{
\item \code{names}: A character vector of the original outcome column names.
\item \code{classes}: A named list of the original outcome classes, or \code{NULL}.
\item \code{levels}: A named list of the original outcome levels for any factor
columns, or \code{NULL}.
}
}

\code{info} is set at \code{\link[=mold]{mold()}} time, and is used in \code{\link[=forge]{forge()}} to validate
\code{new_data}.}

\item{...}{Name-value pairs for additional elements of engines that
subclass this engine.}

\item{subclass}{A character vector. The subclasses of this engine.}
}
\value{
An object of class \code{"hardhat_engine"} containing a list of the supplied
named arguments.
}
\description{
This is the base class for a preprocessing engine. All other engines
subclass this one.
}
