% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-weights.R
\name{new_frequency_weights}
\alias{new_frequency_weights}
\title{Construct a frequency weights vector}
\usage{
new_frequency_weights(x = integer(), ..., class = character())
}
\arguments{
\item{x}{An integer vector.}

\item{...}{Name-value pairs defining attributes}

\item{class}{Name of subclass.}
}
\value{
A new frequency weights vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{new_frequency_weights()} is a developer oriented function for constructing
a new frequency weights vector. Generally, you should use
\code{\link[=frequency_weights]{frequency_weights()}} instead.
}
\examples{
new_frequency_weights()
new_frequency_weights(1:5)
}
