% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-weights.R
\name{new_importance_weights}
\alias{new_importance_weights}
\title{Construct an importance weights vector}
\usage{
new_importance_weights(x = double(), ..., class = character())
}
\arguments{
\item{x}{A double vector.}

\item{...}{Name-value pairs defining attributes}

\item{class}{Name of subclass.}
}
\value{
A new importance weights vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{new_importance_weights()} is a developer oriented function for constructing
a new importance weights vector. Generally, you should use
\code{\link[=importance_weights]{importance_weights()}} instead.
}
\details{
Frequency weights are integers that denote how many times a particular row of
data has been observed. They help compress redundant rows into a single entry.

In tidymodels, frequency weights are used for all parts of the preprocessing,
model fitting, and performance estimation operations.
}
\examples{
new_importance_weights()
new_importance_weights(c(1.5, 2.3, 10))
}
