% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{new_preprocessor}
\alias{new_preprocessor}
\title{Create a new preprocessor}
\usage{
new_preprocessor(engine = new_default_preprocessor_engine(),
  intercept = FALSE, info = info_lst(), ..., subclass = character())
}
\arguments{
\item{engine}{A preprocessing engine. This can be a
\code{"default_preprocessor_engine"}, a \code{"recipe"}, or a \code{"terms"} object.}

\item{intercept}{A logical. Should an intercept be included?}

\item{info}{A named list with 2 elements:
\itemize{
\item \code{"predictors"}: A named list with the following 3 elements:
\itemize{
\item \code{"names"}: A character vector of the original predictor column names.
\item \code{"classes"}: A named list of the original predictor classes, or \code{NULL}.
\item \code{"levels"}: A named list of the original predictor levels for any factor
columns, or \code{NULL}.
}
\item \code{"outcomes"}: A named list with the following 3 elements:
\itemize{
\item \code{"names"}: A character vector of the original outcome column names.
\item \code{"classes"}: A named list of the original outcome classes, or \code{NULL}.
\item \code{"levels"}: A named list of the original outcome levels for any factor
columns, or \code{NULL}.
}
}}

\item{...}{Name-value pairs of extra elements that should be added to
subclassed preprocessors.}

\item{subclass}{An optional character. The subclass of the \code{"preprocessor"}.}
}
\description{
A preprocessor holds a preprocessing engine, and various elements that are
useful for performing structural validation of \code{new_data} in \code{forge()} when
it is time to make predictions using a model.
}
\keyword{internal}
