% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-recipe.R
\name{new_recipe_engine}
\alias{new_recipe_engine}
\title{Create a new recipe engine}
\usage{
new_recipe_engine(mold, forge, intercept = FALSE, info = NULL,
  recipe = NULL, ..., subclass = character())
}
\arguments{
\item{mold}{A named list with two elements, \code{clean} and \code{process}, see
the \code{\link[=new_engine]{new_engine()}} section, Mold Functions, for details.}

\item{forge}{A named list with two elements, \code{clean} and \code{process}, see
the \code{\link[=new_engine]{new_engine()}} section, Forge Functions, for details.}

\item{intercept}{A logical. Should an intercept be included in the
processed data? This information is used by the \code{process} function
in the \code{mold} and \code{forge} function list.}

\item{info}{Either \code{NULL}, or a named list with 2 elements, \code{predictors}
and \code{outcomes}. \code{info} is generated automatically at \code{\link[=mold]{mold()}} time and
is used to validate \code{new_data} at prediction time. The
information in \code{predictors} is the \code{predictors$info} element from running
\code{engine$mold$process()} and the information is \code{outcomes} is the
corresponding \code{outcomes$info}.}

\item{recipe}{Either \code{NULL}, or an unprepped recipe. This argument is set
automatically at \code{\link[=mold]{mold()}} time.}

\item{...}{Name-value pairs for additional elements of engines that
subclass this engine.}

\item{subclass}{A character vector. The subclasses of this engine.}
}
\value{
A preprocessing engine with the class, \code{"recipe_engine"}.
}
\description{
This is the base class for a recipe preprocessing engine. If you want to
create a new preprocessing engine that accepts input from the
user where \code{x} is a recipe, start here.
}
