% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{new_terms_preprocessor_engine}
\alias{new_terms_preprocessor_engine}
\title{Create a terms preprocessor engine}
\usage{
new_terms_preprocessor_engine(predictors, outcomes)
}
\arguments{
\item{predictors}{A terms object for the predictors. The formula should be
for \code{~ RHS}.}

\item{outcomes}{A terms object for the outcomes. The formula should be for
\code{~ LHS}.}
}
\description{
A terms preprocessor engine is composed of two terms objects. One for the
RHS of the original formula, which is used to preprocess the predictors.
And one for the LHS of the original formula, reformulated as \code{~ LHS}.
}
\details{
This is done so that multivariate outcomes can be specified as
\code{Sepal.Width + Sepal.Length ~ Species} and can be processed by
\code{model.frame()} in a meaningful way.
}
\keyword{internal}
