% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-predictor-names.R
\name{predictor-names}
\alias{predictor-names}
\alias{.get_predictor_names_actual}
\alias{.get_predictor_names_original}
\alias{.get_predictor_names_actual.workflow}
\alias{.get_predictor_names_original.workflow}
\alias{.get_predictor_names_actual.recipe}
\alias{.get_predictor_names_original.recipe}
\alias{.get_predictor_names_actual.formula}
\alias{.get_predictor_names_original.formula}
\alias{.get_predictor_names_original.hardhat_blueprint}
\title{Get names of predictors}
\usage{
.get_predictor_names_actual(x, ...)

.get_predictor_names_original(x, ...)

\method{.get_predictor_names_actual}{workflow}(x, include_intercept = FALSE, ...)

\method{.get_predictor_names_original}{workflow}(x, ...)

\method{.get_predictor_names_actual}{recipe}(x, ...)

\method{.get_predictor_names_original}{recipe}(x, ...)

\method{.get_predictor_names_actual}{formula}(x, include_intercept = FALSE, data, ...)

\method{.get_predictor_names_original}{formula}(x, data, ...)

\method{.get_predictor_names_original}{hardhat_blueprint}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Only used in formula methods and are passed to \code{\link[stats:terms]{stats::terms()}}
for original names and \code{\link[stats:model.matrix]{stats::model.matrix()}} for actual names.}

\item{include_intercept}{A logical: should any intercept column be included?}

\item{data}{Original data to be used in a model.}
}
\value{
A sorted character string of column names.
}
\description{
These methods can be used to compute the column names for data that were
given to a function (i.e., "original predictors") or columns that result
after a function has been executed ("actual" predictors).
}
