% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\alias{prepare.data.frame}
\alias{prepare.matrix}
\alias{prepare.formula}
\alias{prepare.recipe}
\title{Prepare data for modeling}
\usage{
prepare(x, ...)

\method{prepare}{data.frame}(x, y, intercept = FALSE, type = "tibble",
  ...)

\method{prepare}{matrix}(x, y, intercept = FALSE, type = "tibble", ...)

\method{prepare}{formula}(formula, data, intercept = FALSE,
  type = "tibble", ...)

\method{prepare}{recipe}(x, data, intercept = FALSE, type = "tibble",
  ...)
}
\arguments{
\item{x}{A data frame, matrix, \code{\link[recipes:recipe]{recipes::recipe()}}, or formula.}

\item{...}{Currently unused.}

\item{y}{A data frame, matrix, or vector depending on \code{x}.}

\item{intercept}{A logical specifying whether or not to include an intercept.}

\item{type}{A single character. One of \code{"tibble"}, \code{"data.frame"}, or
\code{"matrix"} specifying the result type of the predictors.}

\item{data}{A data frame.}
}
\value{
A list containing the \code{predictors}, the \code{outcome}, and the \code{preprocessor}.
}
\description{
\code{prepare()} applies the appropriate processing steps required to get training
data ready to be fed into a model.
\itemize{
\item For a formula method, this applies \code{model.frame()} and \code{model.matrix()}.
\item For a recipe, this performs a call to both \code{\link[recipes:prep]{recipes::prep()}}
and \code{\link[recipes:juice]{recipes::juice()}}.
\item For a data frame or matrix, this uses the \code{default_preprocessor()} which
converts the input to \code{type} and adds an intercept column if requested.
}
}
