% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess.default_preprocessor}
\alias{preprocess.recipes_preprocessor}
\alias{preprocess.terms_preprocessor}
\title{Internal preprocessing of new data}
\usage{
preprocess(preprocessor, new_data, ...)

\method{preprocess}{default_preprocessor}(preprocessor, new_data, ...)

\method{preprocess}{recipes_preprocessor}(preprocessor, new_data,
  outcome = FALSE, ...)

\method{preprocess}{terms_preprocessor}(preprocessor, new_data,
  outcome = FALSE, ...)
}
\arguments{
\item{preprocessor}{A valid \code{"preprocessor"}. The preprocessor that should
be used here is the one in the output from the corresponding call
to \code{\link[=prepare]{prepare()}}.}

\item{new_data}{A data frame or matrix to preprocess.}

\item{...}{Not currently used.}

\item{outcome}{A logical. Should the outcome be processed and returned
as well?}
}
\value{
A named list with elements:
\itemize{
\item \code{predictors}: An object of class \code{type} containing the preprocessed
\code{new_data} predictors.
\item \code{outcome}: If \code{outcome = TRUE}, the outcome is returned here, otherwise
\code{NULL}. If a formula engine was used, this is a data frame that is the
result of extracting the outcome columns from \code{model.frame()}. If a recipe
was used, this is a data.frame that is the result of calling
\code{\link[recipes:bake]{recipes::bake()}} with \code{\link[recipes:all_outcomes]{recipes::all_outcomes()}} specified.
}
}
\description{
\code{preprocess()} applies the transformations requested by the \code{preprocessor}
on a set of \code{new_data} to be used in predictions.
}
\details{
If the outcome is present in \code{new_data}, it can optionally be processed
and returned in the \code{outcome} slot of the returned list. This is only
applicable for the formula and recipes engines, but is very useful when
doing cross validation where you need to preprocess the outcome of a test
set before computing performance.
}
