% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{preprocessor_type}
\alias{preprocessor_type}
\title{Preprocessor printing}
\usage{
preprocessor_type(x)
}
\arguments{
\item{x}{A preprocessor.}
}
\description{
This simply prints out a length 1 character vector specifying the type of
preprocessor.
}
