% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-pred.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vec_ptype_abbr}
\alias{vec_ptype_full}
\alias{obj_print_footer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{vctrs}{\code{\link[vctrs:obj_print]{obj_print_footer}}, \code{\link[vctrs:vec_ptype_full]{vec_ptype_abbr}}, \code{\link[vctrs]{vec_ptype_full}}}
}}

