% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retype.R
\name{retype}
\alias{retype}
\title{Coerce predictors to a new type}
\usage{
retype(x, type = "tibble")
}
\arguments{
\item{x}{Predictors. These are a data frame or a matrix.}

\item{type}{A single character. One of \code{"tibble"}, \code{"data.frame"}, or
\code{"matrix"}.}
}
\description{
\code{retype()} is a simple function that coerces the predictors to a tibble,
data.frame or a matrix.
}
\details{
\code{retype()} is used from \code{\link[=mold]{mold()}} and \code{\link[=forge]{forge()}}.
}
\keyword{internal}
