% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate-quantile_pred.R
\name{snap}
\alias{snap}
\title{Restrict numeric data to the interval [lower, upper]}
\usage{
snap(x, lower, upper, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{lower}{number, the lower bound}

\item{upper}{number, the upper bound}

\item{...}{unused}
}
\value{
An object of the same type as \code{x}
}
\description{
Restrict numeric data to the interval [lower, upper]
}
\keyword{internal}
