% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{update_engine}
\alias{update_engine}
\title{Update a preprocessing engine}
\usage{
update_engine(engine, ...)
}
\arguments{
\item{engine}{A preprocessing engine.}

\item{...}{Name-value pairs of \emph{existing} elements in \code{engine} that should
be updated.}
}
\description{
\code{update_engine()} is the correct way to alter elements of an existing
\code{engine} object. It has two benefits over just doing
\code{engine$elem <- new_elem}.
\itemize{
\item The name you are updating \emph{must} already exist in the engine. This prevents
you from accidentally updating non-existant elements.
\item The constructor for the engine is automatically run after the update by
\code{refresh_engine()} to ensure that the engine is still valid.
}
}
\examples{

engine <- default_xy_engine()

# `intercept` defaults to FALSE
engine

update_engine(engine, intercept = TRUE)

# Can't update non-existant elements
\dontrun{
update_engine(engine, intercpt = TRUE)
}

# Can't add non-valid elements
\dontrun{
update_engine(engine, intercept = 1)
}

}
