% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_fields}
\alias{update_fields}
\title{Update the fields of an existing model}
\usage{
update_fields(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Named fields to update. These names must already exist in \code{x}.}
}
\description{
\code{update_fields()} is a generic function for altering the existing fields
of an existing model object. The default method should be enough to perform
the update, but an  S3 method for your model might be necessary to ensure
that the \emph{types} of the new fields are valid.
}
\details{
\code{update_fields()} is particularly useful in the formula and recipes methods
of the model fit function. There, a preprocessor can be added to the
\code{custom_model} you get back from calling the core fit engine.
}
\keyword{internal}
