% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{utilities}
\alias{utilities}
\alias{glubort}
\alias{extract_terms}
\alias{delete_response}
\alias{dispatchify}
\alias{x_levels}
\title{Utilites}
\usage{
glubort(..., .sep = "", .envir = parent.frame())

extract_terms(x, data)

delete_response(x)

dispatchify(x)

x_levels(terms)
}
\description{
Helper functions that are used internally.
}
\keyword{internal}
