% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_columns_have_type}
\alias{validate_columns_have_type}
\title{Ensure that columns all have the same type}
\usage{
validate_columns_have_type(data, ptype, arg = "outcomes")
}
\arguments{
\item{data}{A tibble to check the columns of.}

\item{ptype}{The type that each column of \code{outcomes} or \code{predictors} should
have. Common types are \code{numeric()} and \code{factor()}.}

\item{arg}{A string. Used in error messaging output.}
}
\description{
\code{validate_columns_have_type()} allows you to supply a \code{ptype} that all
columns should match, otherwise an error
is thrown. It is useful for checking that the user supplied all
numeric predictors, or all factor outcomes.
}
\examples{

# All good
validate_columns_have_type(mtcars, numeric(), "predictors")


# species <- iris[, "Species", drop = FALSE]
# validate_columns_have_type(species, factor())

}
