% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_outcome_is_univariate}
\alias{validate_outcome_is_univariate}
\title{Ensure that the outcome is univariate}
\usage{
validate_outcome_is_univariate(outcomes)
}
\arguments{
\item{outcomes}{A data frame with 1 column.}
}
\description{
The following validation checks are performed:
\itemize{
\item \code{outcomes} must be a data frame
\item \code{outcomes} must have 1 column
}
}
\details{
The easiest way to use this validation function is to supply it the
\code{$outcomes} element of the result of \code{\link[=mold]{mold()}}.
}
\examples{

validate_outcome_is_univariate(data.frame(x = 1))

\dontrun{
validate_outcome_is_univariate(mtcars)
}

}
\concept{validation functions}
