% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_outcomes_is_univariate}
\alias{validate_outcomes_is_univariate}
\alias{check_outcomes_is_univariate}
\title{Ensure that the outcome is univariate}
\usage{
validate_outcomes_is_univariate(outcomes)

check_outcomes_is_univariate(outcomes)
}
\arguments{
\item{outcomes}{Ideally, a data frame with 1 column.}
}
\description{
validate - asserts the following:
\itemize{
\item \code{outcomes} must have 1 column.
}

check - returns the following:
\itemize{
\item \code{ok} A logical. Does the check pass?
\item \code{n_cols} A single numeric. The number of columns.
}
}
\details{
The easiest way to use this validation function is to supply it the
\code{$outcomes} element of the result of \code{\link[=mold]{mold()}}.
}
\examples{

validate_outcomes_is_univariate(data.frame(x = 1))

\dontrun{
validate_outcomes_is_univariate(mtcars)
}

}
\concept{validation functions}
