% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-xy-default.R, R/mold.R
\name{default_xy_engine}
\alias{default_xy_engine}
\alias{mold.data.frame}
\alias{mold.matrix}
\title{Default XY engine}
\usage{
default_xy_engine(intercept = FALSE)

\method{mold}{data.frame}(x, y, engine = NULL, ...)

\method{mold}{matrix}(x, y, engine = NULL, ...)
}
\arguments{
\item{intercept}{A logical. Should an intercept be included in the
processed data? This information is used by the \code{process} function
in the \code{mold} and \code{forge} function list.}

\item{x}{A data frame or matrix containing the predictors.}

\item{y}{A data frame, matrix, or vector containing the outcomes.}

\item{engine}{A preprocessing \code{engine}. If left as \code{NULL}, then a
\code{\link[=default_xy_engine]{default_xy_engine()}} is used.}

\item{...}{Not used.}
}
\description{
This pages holds the details for the XY preprocessing engine. This
is the engine used by default from \code{mold()} if \code{x} and \code{y} are provided
separately (i.e. the XY interface is used).
}
\details{
As documented in \code{\link[=standardize]{standardize()}}, if \code{y} is a \emph{vector}, then the returned
outcomes tibble has 1 column with a standardized name of \code{".outcome"}.

The one special thing about the XY method's forge function is the behavior of
\code{outcomes = TRUE} when a \emph{vector} \code{y} value was provided to the original
call to \code{\link[=mold]{mold()}}. In that case, \code{mold()} converts \code{y} into a tibble, with
a default name of \code{.outcome}. This is the column that \code{forge()} will look
for in \code{new_data} to preprocess. See the examples section for a
demonstration of this.
}
\section{Mold}{


When \code{mold()} is used with the default xy engine:
\itemize{
\item It converts \code{x} to a tibble.
\item It adds an intercept column to \code{x} if \code{intercept = TRUE}.
\item It runs \code{\link[=standardize]{standardize()}} on \code{y}.
}
}

\section{Forge}{


When \code{forge()} is used with the default xy engine:
\itemize{
\item It calls \code{\link[=shrink]{shrink()}} to trim \code{new_data} to only the required columns and
coerce \code{new_data} to a tibble.
\item It calls \code{\link[=scream]{scream()}} to perform validation on the structure of the columns
of \code{new_data}.
\item It adds an intercept column onto \code{new_data} if \code{intercept = TRUE}.
}
}

\examples{
# ---------------------------------------------------------------------------
# Setup

train <- iris[1:100,]
test <- iris[101:150,]

train_x <- train[, "Sepal.Length", drop = FALSE]
train_y <- train[, "Species", drop = FALSE]

test_x <- test[, "Sepal.Length", drop = FALSE]
test_y <- test[, "Species", drop = FALSE]

# ---------------------------------------------------------------------------
# XY Example

# First, call mold() with the training data
processed <- mold(train_x, train_y)

# Then, call forge() with the engine and the test data
# to have it preprocess the test data in the same way
forge(test_x, processed$engine)

# ---------------------------------------------------------------------------
# Intercept

processed <- mold(train_x, train_y, default_xy_engine(intercept = TRUE))

forge(test_x, processed$engine)

# ---------------------------------------------------------------------------
# XY Method and forge(outcomes = TRUE)

# You can request that the new outcome columns are preprocessed as well, but
# they have to be present in `new_data`!

processed <- mold(train_x, train_y)

# Can't do this!
\dontrun{
forge(test_x, processed$engine, outcomes = TRUE)
}

# Need to use the full test set, including `y`
forge(test, processed$engine, outcomes = TRUE)

# With the XY method, if the Y value used in `mold()` is a vector,
# then a column name of `.outcome` is automatically generated.
# This name is what forge() looks for in `new_data`.

# Y is a vector!
y_vec <- train_y$Species

processed_vec <- mold(train_x, y_vec)

# This throws an informative error that tell you
# to include an `".outcome"` column in `new_data`.
\dontrun{
forge(iris, processed_vec$engine, outcomes = TRUE)
}

test2 <- test
test2$.outcome <- test2$Species
test2$Species <- NULL

# This works, and returns a tibble in the $outcomes slot
forge(test2, processed_vec$engine, outcomes = TRUE)

}
