% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce.R
\name{enforce_new_data_level_recovery}
\alias{enforce_new_data_level_recovery}
\title{Recover original factor levels}
\usage{
enforce_new_data_level_recovery(new_data, original_levels)
}
\arguments{
\item{new_data}{A data frame of new predictors and maybe outcomes.}

\item{original_levels}{A named list of the original levels of either the
outcomes or predictors. These are stored in the preprocessor slots
\code{$predictors$levels} and \code{$outcomes$levels}.}
}
\description{
A factor column of \code{new_data} might have a subset of the original factor
levels for some reason. This function checks for that, and recodes the
new factor to have the original levels, with a warning.
}
\details{
Note that \emph{new} factor levels have already been taken care of by now, so
\code{new_data} factor columns can only have either exactly the right levels, or
a subset of the correct levels.
}
\keyword{internal}
