% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{extract_info}
\alias{extract_info}
\title{Extract class and type information}
\usage{
extract_info(x)
}
\arguments{
\item{x}{A data frame or matrix.}
}
\value{
A 0 row slice of \code{x} after converting it to a tibble.
}
\description{
\code{extract_info()} extracts a tibble with 0 rows from \code{x}. This contains
all of the required information about column names, classes, and factor
levels that are required to check the structure of new data at prediction
time.
}
\details{
\code{extract_info()} is useful when creating a new preprocessing \code{engine}. It
extracts the required information that will be used by the validation checks
at prediction time.
}
\examples{

hardhat:::extract_info(iris)

}
\keyword{internal}
