% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-original.R
\name{extract_original_predictors}
\alias{extract_original_predictors}
\title{Extract the original predictors from a model object}
\usage{
extract_original_predictors(preprocessor)
}
\arguments{
\item{preprocessor}{A \code{"preprocessor"} object.}
}
\description{
Often, when \code{predict()}-ing on \code{new_data}, it is a requirement that the
columns in \code{new_data} match the columns in the original data used to fit
the model. \code{extract_original_predictors()} provides a way to extract the
names of those required predictors.
}
\details{
This function knows how to extract 3 different types of predictors, depending
on the type of preprocessing:
\enumerate{
\item Default preprocessing - The original predictors are the same as the
column names stored in \code{preprocessor$predictors}.
\item Formula preprocessing - The original predictors are stored in the
modified \code{terms} attribute \code{"predictors"}. If \code{prepare()} was used on
the input data, then this attribute is constructed for you.
\item Recipe preprocessing - The original predictors are stored in the
\code{variable} column of \code{recipe$var_info}, after filtering down to
\code{type == "predictor"}.
}

This function is called from \code{\link[=shrink]{shrink()}}.
}
