% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-levels.R
\name{extract_outcome_levels}
\alias{extract_outcome_levels}
\title{Extract the levels from a factor outcome}
\usage{
extract_outcome_levels(y, mode, variateness)
}
\arguments{
\item{y}{The outcome. A factor or data frame of factors.}

\item{mode}{A character. Generally either \code{"classification"}
or \code{"regression"}.}

\item{variateness}{A character. One of \code{"univariate"} or \code{"multivariate"}.}
}
\description{
If \code{y} is a factor, or a data frame of factors, then
\code{extract_outcome_levels()} will return the levels of the factor
as either a character vector, or a list of character vectors if
\code{y} is multivariate.
}
