% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-weights.R
\name{is_importance_weights}
\alias{is_importance_weights}
\title{Is \code{x} an importance weights vector?}
\usage{
is_importance_weights(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{is_importance_weights()} checks if \code{x} inherits from
\code{"hardhat_importance_weights"}.
}
\examples{
is_importance_weights(1)
is_importance_weights(frequency_weights(1))
is_importance_weights(importance_weights(1))
}
