% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{is_preprocessor}
\alias{is_preprocessor}
\title{Is \code{x} a valid preprocessor?}
\usage{
is_preprocessor(x)
}
\arguments{
\item{x}{A \code{"recipe"}, \code{"terms"} or \code{"default_preprocessor"} object.}
}
\description{
This function checks to see if \code{x} is a valid preprocessor.
}
\examples{

is_preprocessor(default_preprocessor())

}
