% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mold.R
\name{mold}
\alias{mold}
\alias{mold.data.frame}
\alias{mold.matrix}
\alias{mold.formula}
\alias{mold.recipe}
\title{Mold data for modeling}
\usage{
mold(x, ...)

\method{mold}{data.frame}(x, y, intercept = FALSE, type = "tibble",
  ...)

\method{mold}{matrix}(x, y, intercept = FALSE, type = "tibble", ...)

\method{mold}{formula}(formula, data, intercept = FALSE,
  type = "tibble", ...)

\method{mold}{recipe}(x, data, intercept = FALSE, type = "tibble", ...)
}
\arguments{
\item{x}{A data frame, matrix, or \code{\link[recipes:recipe]{recipes::recipe()}}. If this is a
data.frame or matrix, it should contain the predictors.}

\item{...}{Currently unused.}

\item{y}{A data frame, matrix, or vector containing the outcome(s).}

\item{intercept}{A single logical specifying whether or not to
include an intercept in the molded predictors.}

\item{type}{A single character. One of \code{"tibble"}, \code{"data.frame"}, or
\code{"matrix"} specifying the result type of the predictors.}

\item{formula}{A formula specifying the terms in the format of
\code{outcome ~ predictors}.}

\item{data}{A data frame to mold.}
}
\value{
A named list containing:
\itemize{
\item \code{predictors}: An object of class \code{type} containing the molded predictors
to be used in the model.
\item \code{outcome}: A tibble. If \code{y} was supplied, it is returned after a call to
\code{standardize()} is made. If a formula engine was used, this is a data frame
that is the result of extracting the outcome columns from \code{model.frame()}.
If a recipe was used, this is a data.frame that is the result of calling
\code{\link[recipes:juice]{recipes::juice()}} with \code{\link[recipes:all_outcomes]{recipes::all_outcomes()}} specified.
\item \code{preprocessor}: A \code{"preprocessor"} object for use when making predictions.
}
}
\description{
\code{mold()} applies the appropriate processing steps required to get training
data ready to be fed into a model.
\itemize{
\item For a formula method, this applies \code{model.frame()} and \code{model.matrix()}.
\item For a recipe, this performs a call to both \code{\link[recipes:prep]{recipes::prep()}}
and \code{\link[recipes:juice]{recipes::juice()}}.
\item For a data frame or matrix, this uses the \code{new_default_preprocessor()}s
which converts the input to \code{type} and adds an intercept column if requested.
}
}
