% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mold.R
\name{mold}
\alias{mold}
\title{Mold data for modeling}
\usage{
mold(x, ...)
}
\arguments{
\item{x}{A data frame, matrix, formula, or \code{\link[recipes:recipe]{recipes::recipe()}}. If this is a
data.frame or matrix, it should contain the predictors.}

\item{...}{Currently unused.}
}
\value{
A named list containing 4 elements:
\itemize{
\item \code{predictors}: A tibble containing the molded predictors to be used in the
model.
\item \code{outcome}: A tibble containing the molded outcomes to be used in the
model.
\item \code{preprocessor}: A \code{"preprocessor"} object for use when making predictions.
\item \code{offset}: A tibble with a single column named \code{".offset"} if an offset
was specified in the formula method. Otherwise, \code{NULL}.
}
}
\description{
\code{mold()} applies the appropriate processing steps required to get training
data ready to be fed into a model.

The return values of each method are all consistent with one another, but the
nuances of exactly what is being done for each method vary enough to warrant
separate help files for each. Click through to each one below:
\itemize{
\item XY Method - \code{\link[=mold.data.frame]{mold.data.frame()}} / \code{\link[=mold.matrix]{mold.matrix()}}
\item Formula Method - \code{\link[=mold.formula]{mold.formula()}}
\item Recipes Method - \code{\link[=mold.recipe]{mold.recipe()}}
}
}
