% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{new_base_model}
\alias{new_base_model}
\title{Constructor for a base model}
\usage{
new_base_model(engine, ..., class)
}
\arguments{
\item{engine}{A preprocessing \code{engine} returned from a call to \code{\link[=mold]{mold()}}.}

\item{...}{Name-value pairs for elements specific to the model defined by
\code{class}.}

\item{class}{A character vector representing the class of the model.}
}
\description{
A \strong{model} is a \emph{scalar object}, as classified in
\href{https://adv-r.hadley.nz/s3.html#object-styles}{Advanced R}. As such, it
takes uniquely named fields in \code{...} and combines them into a list with
a class of \code{class}. This entire object represent a single model.
}
\details{
Because every model should have multiple interfaces, including formula
and \code{recipes} interfaces, all models should have an \code{engine} that
can process new data when \code{predict()} is called. The easiest way to generate
an engine with all of the information required at prediction time is to
use the one that is returned from a call to \code{\link[=mold]{mold()}}.
}
\examples{

new_base_model(
  engine = default_xy_engine(),
  custom_field = "my-field",
  class = "custom_model"
)

}
