% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{new_engine}
\alias{new_engine}
\title{Create a new preprocessing engine}
\usage{
new_engine(mold, forge, intercept = FALSE, info = NULL, ...,
  subclass = character())
}
\arguments{
\item{mold}{A named list with two elements, \code{clean} and \code{process}, see
the \code{\link[=new_engine]{new_engine()}} section, Mold Functions, for details.}

\item{forge}{A named list with two elements, \code{clean} and \code{process}, see
the \code{\link[=new_engine]{new_engine()}} section, Forge Functions, for details.}

\item{intercept}{A logical. Should an intercept be included in the
processed data? This information is used by the \code{process} function
in the \code{mold} and \code{forge} function list.}

\item{info}{Either \code{NULL}, or a named list with 2 elements, \code{predictors}
and \code{outcomes}. \code{info} is generated automatically at \code{\link[=mold]{mold()}} time and
is used to validate \code{new_data} at prediction time. The
information in \code{predictors} is the \code{predictors$info} element from running
\code{engine$mold$process()} and the information is \code{outcomes} is the
corresponding \code{outcomes$info}.}

\item{...}{Name-value pairs for additional elements of engines that
subclass this engine.}

\item{subclass}{A character vector. The subclasses of this engine.}
}
\value{
An object of class \code{"hardhat_engine"} containing a list of the supplied
named arguments.
}
\description{
This is the base class for a preprocessing engine. All other engines
subclass this one.
}
