% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{new_preprocessor}
\alias{new_preprocessor}
\title{Create a new preprocessor}
\usage{
new_preprocessor(engine = new_default_preprocessor_engine(),
  intercept = FALSE, type = "tibble", predictors = character(),
  outcomes = character(), ..., subclass = character())
}
\arguments{
\item{engine}{A preprocessing engine. This can be a
\code{"default_preprocessor_engine"}, a \code{"recipe"}, or a \code{"terms"} object.}

\item{intercept}{A logical. Should an intercept be included?}

\item{type}{A character. One of \code{"tibble"}, \code{"data.frame"} or \code{"matrix"}.
The output type for the predictors.}

\item{subclass}{An optional character. The subclass of the \code{"preprocessor"}.}
}
\description{
A preprocessor holds a preprocessing engine, an indicator for whether or
not to include the intercept, and a character string representing the
output type of the predictors.
}
