% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{new_preprocessor}
\alias{new_preprocessor}
\title{Create a new preprocessor}
\usage{
new_preprocessor(engine = new_default_preprocessor_engine(),
  intercept = FALSE, type = "tibble", predictors = character(),
  outcomes = character(), predictor_levels = NULL,
  outcome_levels = NULL, predictor_classes = NULL,
  outcome_classes = NULL, ..., subclass = character())
}
\arguments{
\item{engine}{A preprocessing engine. This can be a
\code{"default_preprocessor_engine"}, a \code{"recipe"}, or a \code{"terms"} object.}

\item{intercept}{A logical. Should an intercept be included?}

\item{type}{A character. One of \code{"tibble"}, \code{"data.frame"} or \code{"matrix"}.
The output type for the predictors.}

\item{predictors}{A character vector. The original predictors used when
fitting the model.}

\item{outcomes}{A character vector. The original outcome columns used
when fitting the model.}

\item{predictor_levels}{A named list. Each element of the list should be a
character vector of levels. The names of the list should be the \emph{predictor}
factor columns in the training data. If there are no factor columns, this is
\code{NULL}.}

\item{outcome_levels}{A named list. Each element of the list should be a
character vector of levels. The names of the list should be the \emph{outcome}
factor columns in the training data. If there are no factor columns, this is
\code{NULL}.}

\item{predictor_classes}{A named list. Each element of the list should
be a character vector of classes. The names of the list are the \emph{predictor}
names in the training data.}

\item{outcome_classes}{A named list. Each element of the list should
be a character vector of classes. The names of the list are the \emph{outcome}
names in the training data.}

\item{...}{Name-value pairs of extra elements that should be added to
subclassed preprocessors.}

\item{subclass}{An optional character. The subclass of the \code{"preprocessor"}.}
}
\description{
A preprocessor holds a preprocessing engine, an indicator for whether or
not to include the intercept, and a character string representing the
output type of the predictors.
}
\keyword{internal}
