% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\alias{prepare.data.frame}
\alias{prepare.matrix}
\alias{prepare.formula}
\alias{prepare.recipe}
\title{Prepare data for modeling}
\usage{
prepare(x, ...)

\method{prepare}{data.frame}(x, y, intercept = FALSE, type = "tibble",
  ...)

\method{prepare}{matrix}(x, y, intercept = FALSE, type = "tibble", ...)

\method{prepare}{formula}(formula, data, intercept = FALSE,
  type = "tibble", ...)

\method{prepare}{recipe}(x, data, intercept = FALSE, type = "tibble",
  ...)
}
\arguments{
\item{x}{A data frame, matrix, or \code{\link[recipes:recipe]{recipes::recipe()}}. If this is a
data.frame or matrix, it should contain the predictors.}

\item{...}{Currently unused.}

\item{y}{A data frame, matrix, or vector containing the outcome(s).}

\item{intercept}{A single logical specifying whether or not to
include an intercept in the prepared predictors.}

\item{type}{A single character. One of \code{"tibble"}, \code{"data.frame"}, or
\code{"matrix"} specifying the result type of the predictors.}

\item{formula}{A formula specifying the terms in the format of
\code{outcome ~ predictors}.}

\item{data}{A data frame to prepare.}
}
\value{
A named list containing:
\itemize{
\item \code{predictors}: An object of class \code{type} containing the prepared predictors
to be used in the model.
\item \code{outcome}: If \code{y} was supplied, it is returned unmodified here. If a
formula was used, this is the result of \code{\link[=model.response]{model.response()}}. If a recipe
was used, this is a data.frame that is the result of calling
\code{\link[recipes:juice]{recipes::juice()}} with \code{\link[recipes:all_outcomes]{recipes::all_outcomes()}} specified.
\item \code{preprocessor}: A \code{"preprocessor"} object for use when making predictions.
}
}
\description{
\code{prepare()} applies the appropriate processing steps required to get training
data ready to be fed into a model.
\itemize{
\item For a formula method, this applies \code{model.frame()} and \code{model.matrix()}.
\item For a recipe, this performs a call to both \code{\link[recipes:prep]{recipes::prep()}}
and \code{\link[recipes:juice]{recipes::juice()}}.
\item For a data frame or matrix, this uses the \code{new_default_preprocessor()}s
which converts the input to \code{type} and adds an intercept column if requested.
}
}
