% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Internal preprocessing of new data}
\usage{
preprocess(preprocessor, new_data, intercept = FALSE, type = "tibble")
}
\value{
The preprocessed data, returned as a matrix of numeric columns.
}
\description{
\code{preprocess()} applies the transformations requested by the \code{preprocessor}
on a set of \code{new_data} to be used in predictions.
}
\details{
The \code{preprocessor} can be a \code{recipe::recipes()} object, a \code{terms} object
resulting from the use of a formula method, or a \code{default_preprocessor()}
which converts \code{new_data} to a matrix and optionally adds an intercept.
}
