% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Internal preprocessing of new data}
\usage{
preprocess(preprocessor, new_data)
}
\arguments{
\item{preprocessor}{A valid preprocessor. Can be a \code{\link[recipes:recipe]{recipes::recipe()}}, a
terms object, or the result of \code{default_preprocessor()}.}

\item{new_data}{A data frame to preprocess.}
}
\value{
The preprocessed data, returned as a \code{type}.
}
\description{
\code{preprocess()} applies the transformations requested by the \code{preprocessor}
on a set of \code{new_data} to be used in predictions.
}
\details{
The \code{preprocessor} can be a \code{recipe::recipes()} object, a \code{terms} object
resulting from the use of a formula method, or a \code{default_preprocessor()}
which converts \code{new_data} to a matrix and optionally adds an intercept.
}
