% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Subset only required predictors}
\usage{
shrink(object, new_data)
}
\arguments{
\item{object}{The model fit. If the default preprocessing was supplied, an
\code{object$predictors} element should exist holding a character vector of the
predictor names. Otherwise, as long as \code{prepare()}
was called, then nothing else needs to be done.}

\item{new_data}{A data frame containing the new data that will be shrunk down
to only contain the predictors required for preprocessing and prediction.}
}
\description{
\code{shrink()} first validates that all \code{predictors} exist in \code{new_data}, and
then subsets \code{new_data} to only contain those predictors.
}
\keyword{internal}
