% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Subset only required columns}
\usage{
shrink(preprocessor, new_data, outcome = FALSE)
}
\arguments{
\item{preprocessor}{A \code{"preprocessor"}.}

\item{new_data}{A data frame at least containing the new predictors, and
potentially the outcomes if \code{outcome = TRUE}.}

\item{outcome}{A logical. Should the outcome be included in the shrunk
\code{new_data} output as well?}
}
\value{
A tibble containing the required predictors (and potentially the
outcomes).
}
\description{
\code{shrink()} subsets \code{new_data} to only contain the required columns.
\code{shrink()} also performs the following validation checks on \code{new_data}:
\itemize{
\item Checks that all required predictor columns exist in \code{new_data}.
}
}
\details{
If \code{outcome = TRUE}, then the validation checks are performed on the known
outcome column as well, and it is returned along with the predictors. If
\code{\link[=mold]{mold()}} was called with the XY interface, then no preprocessing was done
to \code{y} and \code{outcome} will have no effect (if a vector was passed as \code{y}
during the fit, \code{shrink()} has no way of knowing what column in \code{new_data}
corresponds to the outcome).

\code{shrink()} is called by \code{\link[=forge]{forge()}} before \code{\link[=scream]{scream()}} and before the actual
processing is done.
}
\examples{
train <- iris[1:100,]
test <- iris[101:150,]

# mold() is run at model fit time
# and a terms preprocessor is recorded
x <- mold(log(Sepal.Width) ~ Species, train)

# Pass that preprocessor to shrink(), along with new_data
# to get a tibble of required predictors back
shrink(x$preprocessor, test)

# outcome = TRUE also returns the outcome,
# this can be useful if doing cross validation
shrink(x$preprocessor, test, outcome = TRUE)

# forge() actually preprocesses the new_data, after
# a call to shrink(). Notice how now the Sepal.Width
# column is logged
forge(x$preprocessor, test, outcome = TRUE)$outcomes

}
