% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.R
\name{spruce}
\alias{spruce}
\alias{spruce.default}
\alias{spruce.response}
\alias{spruce.class}
\alias{spruce.prob}
\title{Spruce up predictions}
\usage{
spruce(type, new_data, ...)

\method{spruce}{default}(type, new_data, ...)

\method{spruce}{response}(type, new_data, .pred, ...)

\method{spruce}{class}(type, new_data, .pred_class, ...)

\method{spruce}{prob}(type, new_data, .pred_levels, .prob_matrix, ...)
}
\arguments{
\item{type}{The return value from \code{dispatchify()} called on a character
\code{type}. This is used for appropriate dispatching.}

\item{new_data}{The post processed new data to make predictions for. This is
used to validate that the number of rows in the predictions is the same
as the number of rows in the original \code{new_data}.}

\item{.pred}{(\code{type = "response"}) A numeric vector of predictions.}

\item{.pred_class}{(\code{type = "class"}) A factor of "hard" class predictions.}

\item{.pred_levels, .prob_matrix}{(\code{type = "prob"}) \code{.pred_levels} should
be a character vector of the original levels of the outcome used in training.
\code{.prob_matrix} should be a numeric matrix of class probabilities with as many
columns as levels in \code{.pred_levels}, and in the same order.}
}
\value{
A tibble with as many rows as in \code{new_data}. The column names and number
of columns vary based on the \code{type}, but are standardized.
}
\description{
\code{spruce()} converts predictions into a standardized format. It is generally
called from your prediction implementation function for the specific
\code{type} of prediction to return.
}
