% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-response.R, R/util.R
\name{delete_response}
\alias{delete_response}
\alias{utilities}
\alias{glubort}
\alias{simplify_terms}
\title{Utilites}
\usage{
delete_response(x)

glubort(..., .sep = "", .envir = parent.frame())

simplify_terms(x)
}
\description{
Helper functions that are used internally.
}
\keyword{internal}
