% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{utilities}
\alias{utilities}
\alias{glubort}
\alias{simplify_terms}
\title{Utilites}
\usage{
glubort(..., .sep = "", .envir = parent.frame())

simplify_terms(x)
}
\description{
Helper functions that are used internally.
}
\keyword{internal}
