% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_prediction_size}
\alias{validate_prediction_size}
\alias{check_prediction_size}
\title{Ensure that predictions have the correct number of rows}
\usage{
validate_prediction_size(.pred, new_data)

check_prediction_size(.pred, new_data)
}
\arguments{
\item{.pred}{A tibble. The predictions to return from any prediction
\code{type}. This is often created using one of the spruce functions, like
\code{\link[=spruce_response]{spruce_response()}}.}

\item{new_data}{A data frame of new predictors and possibly outcomes.}
}
\description{
validate - asserts the following:
\itemize{
\item The number of rows in \code{.pred} must be the same as the number
of rows in \code{new_data}.
\item \code{.pred} must be a data frame.
}

check - returns the following:
\itemize{
\item \code{ok} A logical. Does the check pass?
\item \code{n_new_data} A single numeric. The number of rows in \code{new_data}.
\item \code{n_pred} A single numeric. The number of rows in \code{.pred}.
}
}
\details{
This validation function is one that is more developer focused rather than
user focused. It is a final check to be used right before a value is
returned from your specific \code{predict()} method, and is mainly a "good
practice" sanity check to ensure that your prediction engine always returns
the same number of rows as \code{new_data}, which is one of the modeling
conventions this package tries to promote.
}
\section{Validation}{


hardhat provides validation functions at two levels.
\itemize{
\item \code{check_*()}:  \emph{check a condition, and return a list}. The list
always contains at least one element, \code{ok}, a logical that specifies if the
check passed. Each check also has check specific elements in the returned
list that can be used to construct meaningful error messages.
\item \code{validate_*()}: \emph{check a condition, and error if it does not pass}. These
functions call their corresponding check function, and
then provide a default error message. If you, as a developer, want a
different error message, then call the \code{check_*()} function yourself,
and provide your own validation function.
}
}

\examples{
# Say new_data has 5 rows
new_data <- mtcars[1:5,]

# And somehow you generate predictions
# for those 5 rows
.pred_vec <- 1:5

# Then you use `spruce_response()` to clean
# up these numeric predictions
.pred <- spruce_response(.pred_vec)

.pred

# Use this check to ensure that
# the number of rows or .pred match new_data
check_prediction_size(.pred, new_data)

# An informative error message is thrown
# if the rows are different
\dontrun{
validate_prediction_size(spruce_response(1:4), new_data)
}

}
\seealso{
Other validation functions: \code{\link{validate_new_data_column_names}},
  \code{\link{validate_outcomes_is_univariate}}
}
\concept{validation functions}
