% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\alias{validate_inherits}
\alias{validate_is}
\alias{validate_length}
\alias{validate_mode}
\alias{validate_variateness}
\alias{validate_has_named_columns}
\alias{validate_has_unique_names}
\alias{validate_x_has_numeric_cols}
\alias{validate_y_univariate}
\alias{validate_predictors}
\alias{validate_outcomes}
\alias{validate_recipes_available}
\alias{validate_prediction_size}
\alias{validate_new_data_classes}
\alias{check_new_data_factor_levels}
\alias{validate_predictor_arguments}
\alias{validate_intercept}
\alias{validate_formula_has_intercept}
\title{Validation}
\usage{
validate_inherits(x, x_nm, cls)

validate_is(.x, .f, .expected, .x_nm, .note = "")

validate_length(.x, .x_nm, .n = 1L)

validate_mode(mode)

validate_variateness(variateness)

validate_has_named_columns(x, x_nm)

validate_has_unique_names(x, x_nm)

validate_x_has_numeric_cols(x)

validate_y_univariate(y)

validate_predictors(new_data, predictors)

validate_outcomes(new_data, outcomes)

validate_recipes_available()

validate_prediction_size(.pred, new_data)

validate_new_data_classes(terms, new_model_frame)

check_new_data_factor_levels(x_levels, new_data)

validate_predictor_arguments(x, x_nm = "x", intercept = FALSE)

validate_intercept(intercept)

validate_formula_has_intercept(formula)
}
\description{
Various input validation helpers. These should not be exported.
}
\details{
\code{@keywords internal} is used to generate documentation even though the
function is not exposed to the user. This is helpful for developers as it
allows you to document your internal functions.
}
\keyword{internal}
