% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-internal.R
\name{validation}
\alias{validation}
\alias{validate_is}
\alias{validate_has_unique_names}
\alias{validate_has_unique_column_names}
\alias{validate_x_has_numeric_cols}
\alias{validate_y_univariate}
\alias{validate_predictors}
\alias{validate_outcomes}
\alias{validate_recipes_available}
\alias{validate_prediction_size}
\alias{validate_new_data_classes}
\alias{validate_intercept}
\alias{validate_formula_has_intercept}
\title{Validation}
\usage{
validate_is(.x, .f, .expected, .x_nm, .note = "")

validate_has_unique_names(x, x_nm)

validate_has_unique_column_names(x, x_nm)

validate_x_has_numeric_cols(x)

validate_y_univariate(y)

validate_predictors(new_data, predictors)

validate_outcomes(new_data, outcomes)

validate_recipes_available()

validate_prediction_size(.pred, new_data)

validate_new_data_classes(new_data, original_classes)

validate_intercept(intercept)

validate_formula_has_intercept(formula)
}
\description{
Various input validation helpers. These should not be exported.
}
\details{
\code{@keywords internal} is used to generate documentation even though the
function is not exposed to the user. This is helpful for developers as it
allows you to document your internal functions.
}
\keyword{internal}
